Authentication: fixed a CSRF bug in the refresh token handling.
the previous behavior is sending a new access and refresh token
along with the query response everytime the user sends an expired
access token, this behavior can cause a CSRF since the refresh
token stored in the cookies can function as an access token.

this fix errors out the request if the access token is expired,
forcing the user to request a new access token using the /request_token
route.

also, deez nuts.